Sign up Page and Login page done
