--- foodie/src/Component/Pag.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function pag() {
-  return (
-    <div>page</div>
-  )
-}
-
-export default pag
