+++ foodie/src/Styles/mainpage.module.css
@@ -22,6 +22,8 @@
 input {
     border-radius: 5px;
     background-color: #D9D9D9;
+    width: 70%;
+    padding: 10px;
 }
 ::placeholder {
     color: #000;
